Add Protection Command for Chests Doors etc chest / falldoor / door / fence-gate protection with permissions for player and groups with flags Add Messages for Protection Modify etc May add Views to increase performance (View with only location type 5) Add Missing Up and Down Sign Action Add Sign Edit Add Sign Copy / Paste Add Fake AFK (/admin Command) Add Chat Clear (/admin Command) Add Top (/admin Command) Add Bank Upgrade Add Redstoneless Lights /admin light Add NPC Adventurer, Baker, Butcher, Farmer, Fisher, Florist, Lumberjack, Miner, Musican, Shepherd, Smith, Bag Salesman, Banker, Beekeeper, Enchanter Rewrite NPC Stuff (less hardcoded) Add NPCs to Database Fixed Playerhead Command Fixed Playerhead Event Add Request to Mojang API to get Textures Add CustomHeads (with textures as base64, uuid and name) Add Bags (Miner, Farmer) Database based (softcoded) with insert in bag_type new bag Add Sign Up and Down Command Add Protection Lock Materials / Types in Database Add Protection overwrite by admin and mod group Add interest for bank every day (on login) Add Missing Strings from Admin.java Command Add Missing Strings from Worlds.java Command Included /rellu ping to /admin ping Add Sudo Command to sudo as other Player Add Exit Command to self kick Players. Or Close Server as Console Sender Add NPC Builder who sells Buildblocks Enchantment System with custom enchantments Add Telekinesis Add Thunderstroke Add Replenishment Add Delicate Add Auto Smelting

+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/helpers/pojo/BankAccountEntry.java
@@ -16,7 +16,7 @@ public class BankAccountEntry {
     private String deleted;
     private int deletedBy;
     private int player_fk;
-    private float value;
+    private double value;
     private BankTierEntry tier;
 
     public int getId() {
@@ -83,11 +83,11 @@ public class BankAccountEntry {
         this.player_fk = playerId;
     }
 
-    public float getValue() {
+    public double getValue() {
         return value;
     }
 
-    public void setValue(float value) {
+    public void setValue(double value) {
         this.value = value;
     }
